feat: Tickets are shown on the correct tab
Earlier active tickets that expired was "hanging around" on the active
tab until the view was remounted. With this change the tickets are
rendered on their correct tab, and changing tab just as it expires.
